Stop following login redirect so rapidly
There's code here designed to handle a login-requested event, let's
use it.

Previously, xqueue would respond with a 302 that was immediately
followed (by the requests library under-the-hood) to the /login
URL, but with the exact same parameters. For some reason, this
led to a 200 response containing parseable json, so this code
block had no idea and was unable to fix the problem.

TNL-5928