fix: Switch from imp to importlib.
The import in `grader.py` was just not being used so I dropped it.  In
jailedgrader.py I replaced the implementation with the sourcefile loader
from importlib which is still around.  The `imp` module is deprecated
and was removed in Python 3.12

With these changes the code is compatible with Python 3.8 and 3.12.